mycelium-backend/pyproject.toml
[project]
name = "mycelium-backend"
version = "0.1.0"
description = "Mycelium FastAPI backend — async room coordination via Postgres LISTEN/NOTIFY"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.115.0,<0.116",
    "asyncpg>=0.29.0,<0.32",
    "pydantic-settings>=2.5.2,<3",
    "httpx>=0.28.1",
    "python-multipart>=0.0.9",
    "anthropic>=0.40.0",
    "litellm>=1.60.0",
    "greenlet>=3.0",
    "agensgraph-python==1.0.2",
    "psycopg2-binary>=2.9.0",
    "negmas>=0.10.0",
    "fastembed>=0.4.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0,<2",
]

[dependency-groups]
dev = [
    "pre-commit>=3.4.0,<4",
    "ruff>=0.11.0,<0.12",
    "ty>=0.0.1a8",
    "watchdog>=6.0.0,<7",
    "python-dotenv>=1.0.1,<2",
    "pytest>=9.0.1,<10",
    "pytest-asyncio>=1.3.0,<2",
    "aiosqlite>=0.22.1",
    "respx>=0.22.0",
]

[tool.uv]
package = false

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "S"]
ignore = ["E501", "B008", "B023", "B904", "S110", "S112"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert statements are idiomatic in tests
    "PLR2004", # magic values in assertions are expected
    "ANN",     # type annotations not required in tests
    "ARG001",  # unused function args (e.g. fixtures)
    "SLF001",  # private member access in tests is fine
]

[tool.ty.environment]
python-version = "3.12"
